src/InOneWeekend/requirements.txt
numpy
pillow
xarray
